Ceciliasharonmerge v3

+++ react-app/src/components/EditMessageForm/index.js
@@ -72,20 +72,26 @@ const EditMessageForm = () => {
             onChange={updateContent}
           />
           <div className="UpdateMessageButtonContainer">
-            <div className="DeleteMessageButton">
-              <button
-                onClick={async () => {
-                  await dispatch(deleteMessage(channel_id, id)).then(() =>
-                    dispatch(loadChannel(channel_id))
-                  );
-                }}
-              >
-                Delete
-              </button>
-            </div>
             <div className="UpdateMessageButton">
               <button type="submit" disabled={errors.length > 0}>
                 Update
+              </button>
+            </div>
+            <div className="DeleteMessageButton">
+              <button
+                onClick={async () => {
+                  if (
+                    window.confirm(
+                      "Are you sure you want to delete this message?"
+                    )
+                  ) {
+                    await dispatch(deleteMessage(channel_id, id)).then(() =>
+                      dispatch(loadChannel(channel_id))
+                    );
+                  }
+                }}
+              >
+                Delete
               </button>
             </div>
           </div>
